# openapi-power-automate.yaml
swagger: '2.0'
info:
  title: SLH TMS API
  version: '1.2'
host: slh-tms-api-prod.gentlepond-08dba66b.uksouth.azurecontainerapps.io
basePath: /api/v1
schemes: [https]
produces: [application/json]
consumes: [application/json]
securityDefinitions:
  oauth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://login.microsoftonline.com/REPLACE_TENANT_ID/oauth2/v2.0/authorize
    tokenUrl: https://login.microsoftonline.com/REPLACE_TENANT_ID/oauth2/v2.0/token
    scopes:
      api://REPLACE_API_CLIENT_ID/Tms.Access: Access the TMS API
security: [{ oauth2: [api://REPLACE_API_CLIENT_ID/Tms.Access] }]
paths:
  /customers:
    get:
      operationId: GetCustomers
      summary: Search active customers
      parameters: [{ name: q, in: query, type: string, required: false }]
      responses: { '200': { description: Customer results, schema: { type: array, items: { $ref: '#/definitions/Customer' } } } }
...
  /drivers:
    get:
      operationId: GetDrivers
      summary: Search active drivers
      parameters: [{ name: q, in: query, type: string, required: false }]
      responses: { '200': { description: Driver results, schema: { type: array, items: { $ref: '#/definitions/Driver' } } } }
...
  /sites:
    get:
      operationId: GetSites
      summary: Search active collection and delivery sites
      parameters: [{ name: q, in: query, type: string, required: false }]
      responses: { '200': { description: Site results, schema: { type: array, items: { $ref: '#/definitions/Site' } } } }
...
  /staging:
    get:
      operationId: ListStagedImports
      summary: List staged records for review
      parameters:
        - { name: status, in: query, type: string, required: false, enum: [PendingReview, Approved, Rejected, Promoted, Failed] }
        - { name: entityType, in: query, type: string, required: false }
        - { name: take, in: query, type: integer, required: false, default: 100, maximum: 2000 }
      responses: { '200': { description: Staged records } }
    post:
      operationId: SubmitStagedImport
      summary: Submit an import for review
      parameters:
        - name: body
          in: body
          required: true
          schema: { $ref: '#/definitions/StageImportRequest' }
      responses:
        '202': { description: Accepted for review, schema: { $ref: '#/definitions/StageImportResponse' } }
        '200': { description: Existing idempotent submission, schema: { $ref: '#/definitions/StageImportResponse' } }
...
definitions:
...
  StageImportRequest:
    type: object
    required: [entityType, idempotencyKey, payload]
    properties:
      entityType: { type: string, enum: [customer, vehicle, driver, trailer, site, marketcontact, order] }
      idempotencyKey: { type: string, example: email-message-id-or-source-key }
      source: { type: string, example: PowerAutomate/InfoMailbox }
      payload: { type: object, additionalProperties: true }
  StageImportResponse:
    type: object
    properties:
      stagingId: { type: string, format: uuid }
      status: { type: string, example: PendingReview }
      receivedAtUtc: { type: string, format: date-time }
      reviewUrl: { type: string }
...
  Customer:
    type: object
    properties: { id: { type: string, format: uuid }, code: { type: string }, name: { type: string }, active: { type: boolean } }
...
  Driver:
    type: object
    properties: { id: { type: string, format: uuid }, employeeNumber: { type: string }, displayName: { type: string }, tachoName: { type: string }, driverType: { type: string }, driverGroup: { type: string }, skills: { type: string }, active: { type: boolean } }
...
  Site:
    type: object
    properties: { id: { type: string, format: uuid }, externalCode: { type: string }, name: { type: string }, driverTextName: { type: string }, collectionAddress: { type: string }, collectionInstructions: { type: string }, mapLink: { type: string }, active: { type: boolean } }
...